Update whitelist_slim.txt
Remove *.branch.io

Info found on: https://www.reddit.com/r/pihole/comments/wpu3us/api2branchio/

When a network request is made to ‘api2.branch.io’, the results recorded allow Spotify to verify if features such as the above succeeded or not. For example if a link to the Spotify’s service redirects users to the intended location, or if launching the Spotify app on a device was successful.

In order to achieve this some data is processed and shared with Branch, such as the IP address, referral identification for the request, and actions taken by the user.

The requests to Branch are made once per engagement, such as when opening a link to Spotify or opening the Spotify app. Once the Spotify app has successfully reached Branch and logged the engagement, the connection has succeeded and will stop. However, when the connection attempts to Branch for these engagements are denied, the Spotify app will retry the requests until they succeed as they are read and interpreted as a network connection error.

As a result of your report, we have initiated with Branch the work necessary to prevent blocked requests from being retried, with the goal that the retries are limited when the connection attempts fail.